% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggjoy.R
\docType{package}
\name{ggjoy}
\alias{ggjoy}
\alias{ggjoy-package}
\alias{geom_joy}
\alias{geom_joy2}
\alias{geom_joy_gradient}
\alias{stat_joy}
\alias{theme_joy}
\title{ggjoy}
\description{
This package is deprecated. Please switch over to using \link{ggridges}.
}
